requirements.txt
# Function dependencies, for example:
#package>=version
python-binance>=0.7.9
numpy==1.20.2
pandas==1.2.4
datetime
sklearn
ta
xgboost